Feature/api integration

+++ docs/.vuepress/store/modules/admin/actions.js
@@ -1,22 +1,29 @@
 import AllApplications from '../../../api/admin/allApplications';
 import AllUsers from '../../../api/admin/allUsers';
 import AllPlans from '../../../api/admin/allPlans';
+import ModalWindow from '../../../services/ModalWindow';
 
 export default {
   getAllApplications({ commit }) {
     AllApplications.get()
-      .then(({ data }) => {
-        commit('setAllApplications', data);
+      .then(({ data: { appData } }) => {
+        commit('setAllApplications', appData);
       })
       .catch(console.log);
   },
 
-  deleteApplicationByUser({ dispatch }, { userName, appId }) {
-    AllApplications.deleteById(userName, appId)
-      .then(() => {
-        dispatch('getAllApplications');
-      })
-      .catch(console.log);
+  async deleteApplicationByUser({ dispatch }, { userName, appId }) {
+    try {
+      const confirm = await ModalWindow.openDialog();
+
+      if (confirm) {
+        await AllApplications.deleteById(userName, appId);
+        await dispatch('getAllApplications');
+        await dispatch('application/getApps', null, { root: true });
+      }
+    } catch (error) {
+      console.log(error);
+    }
   },
 
   getAllUsers({ commit }) {
@@ -24,14 +31,17 @@ export default {
       .then(({ data }) => {
         commit('setAllUsers', data.users);
       })
-      .catch(console.log);
+      .catch(console.log)
+      .finally(() => {
+        return Promise.resolve();
+      });
   },
 
   async updateRole({ dispatch, state }, userId) {
     const { role, userName } = state.allUsers.find(user => user.userId === userId);
 
     const roles = [
-      'user', 'admin', 'superadmin'
+      'user', 'admin'
     ];
 
     const roleIndex = roles.indexOf(role) + 1;
@@ -40,42 +50,60 @@ export default {
       role: roles[roleIndex] ? roles[roleIndex] : roles[0]
     };
 
-    AllUsers.updateRole(userName, data)
-      .then(() => {
-        dispatch('getAllUsers');
-      })
-      .catch(console.log)
-      .finally(() => {
-        // TODO request is very quickly, maybe need set promise for getAllUsers
-        return Promise.resolve();
-      });
+    try {
+      const confirm = await ModalWindow.openDialog();
+
+      if (confirm) {
+        await AllUsers.updateRole(userName, data);
+        await dispatch('getAllUsers');
+      }
+    } catch (error) {
+      console.log(error);
+    }
+
+    return Promise.resolve();
   },
 
-  setUserStatus({ dispatch, state }, userName) {
-    const { userEnabled } = state.allUsers.find(user => user.userName === userName);
-    AllUsers.setUserStatus(userName, userEnabled)
-      .then(() => {
-        dispatch('getAllUsers');
-      })
-      .catch(console.log);
+  async setUserStatus({ dispatch, state }, userName) {
+    try {
+      const confirm = await ModalWindow.openDialog();
+
+      if (confirm) {
+        const { userEnabled } = state.allUsers.find(user => user.userName === userName);
+
+        await AllUsers.setUserStatus(userName, userEnabled);
+        await dispatch('getAllUsers');
+      }
+    } catch (error) {
+      console.log(error);
+    }
   },
 
-  deleteUserByUsername({ dispatch }, userName) {
-    AllUsers.deleteByUsername(userName)
-      .then(() => {
-        dispatch('getAllUsers');
-      })
-      .catch(console.log);
+  async deleteUserByUsername({ dispatch }, userName) {
+    try {
+      const confirm = await ModalWindow.openDialog();
+
+      if (confirm) {
+        await AllUsers.deleteByUsername(userName);
+        await dispatch('getAllUsers');
+      }
+    } catch (error) {
+      console.log(error);
+    }
   },
 
-  changePublishedState({ dispatch }, { planId, published }) {
-    AllPlans.updateStateById({
-      planId,
-      published: ! published
-    })
-      .then(() => {
-        dispatch('usagePlans/getUsagePlans', null, { root: true });
-      })
-      .catch(console.log);
+  async changePublishedState({ dispatch }, { planId, published }) {
+    try {
+      const confirm = await ModalWindow.openDialog();
+
+      if (confirm) {
+        await AllPlans.updateStateById({ planId, published: ! published });
+        await dispatch('usagePlans/getUsagePlans', null, { root: true });
+      }
+    } catch (error) {
+      console.log(error);
+    }
+
+    return Promise.resolve();
   }
 };
